requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# telebot/config.py: 1
dynaconf == 3.1.7

# telebot/main.py: 20
python_telegram_bot == 13.7
